# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ABSuite API
  version: 1.0.0
  description: |
    Scoped, expiring, auditable credentials and execution infrastructure for AI agents.

    ## Authentication

    Three credential types, used for different purposes:

    | Header | Purpose |
    |---|---|
    | `Authorization: Bearer <capability-token>` | Normal service calls. Must carry the scope the endpoint requires. |
    | `X-ABSuite-Admin-Key` | Bootstrap credential. Grants full authority; used to mint the first token and manage tenants. |
    | `X-ABSuite-Tenant-Key` | Identifies the billed tenant. Optional — omit for an unmetered self-hosted deployment. |

    ## Scopes

    Scopes are `resource:action`, matched segment-wise. `read:*` grants
    `read:users` but never `read:users:delete`. A bare `*` grants everything.

    ## Quotas

    When a tenant key is supplied, plan limits are enforced. Exceeding a limit
    returns `402`; a suspended account returns `403`.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: CapKit
  - url: http://localhost:8082
    description: Edge-Run
  - url: http://localhost:8083
    description: QuickBench
  - url: http://localhost:8084
    description: Connector-Starter
...
components:
  securitySchemes:
    capabilityToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A CapKit capability token carrying the required scope.
    adminKey:
      type: apiKey
      in: header
      name: X-ABSuite-Admin-Key
    tenantKey:
      type: apiKey
      in: header
      name: X-ABSuite-Tenant-Key
...
paths:
...
  /run:
    post:
      tags: [Benchmarking]
      summary: Start a benchmark
      description: Requires scope `bench:run`. QuickBench, port 8083.
      security:
        - capabilityToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider]
              properties:
                provider: { type: string, enum: [ollama, openai, anthropic, http] }
                model: { type: string }
                url: { type: string, description: Required for the http provider }
                name: { type: string }
                testRuns: { type: integer, default: 10, maximum: 500 }
                warmupRuns: { type: integer, default: 2 }
                concurrency: { type: integer, default: 1, maximum: 32 }
      responses:
        '202':
          description: Accepted; poll the job for results
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string }
                  status: { type: string }
                  estimatedDuration: { type: string }
  /compare:
    get:
      tags: [Benchmarking]
      summary: Compare two benchmark runs
      description: |
        Requires scope `bench:read`. Uses Welch's t-test, which does not assume
        equal variance. Suitable for gating a deploy.
      security:
        - capabilityToken: []
      parameters:
        - { name: baseline, in: query, required: true, schema: { type: string } }
        - { name: candidate, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: Comparison verdict
          content:
            application/json:
              schema:
                type: object
                properties:
                  deltaMeanMs: { type: number }
                  deltaPercent: { type: number }
                  significant: { type: boolean }
                  verdict: { type: string, enum: [regression, improvement, 'no significant change'] }
...